Update multigrainOutputParser.py
I found the problem in the multigrainOutputParser. When parsing the GrainSpotter log-file the summary lines at the end were deleted to get the pure grain information. But when saving the information into a new log-file the summary lines were not added again. So when repeating the parsing and saving several times, more and more rows at the end of the file were deleted until the script threw an error.

However, I haven't found the place where to get the info about the g-vectors (how many in total, how many assigned, how many assigned twice etc.) For now, this is not important because the changes I made are already enough to make the script working but it has still to be implemented at some point.